+++ photutils/psf/functional_models.py
@@ -8,7 +8,6 @@ import numpy as np
 from astropy.modeling import Fittable2DModel, Parameter
 from astropy.modeling.utils import ellipse_extent
 from astropy.units import UnitsError
-from astropy.utils.decorators import deprecated
 from scipy.special import erf, j1, jn_zeros
 
 __all__ = [
@@ -18,7 +17,6 @@ __all__ = [
     'CircularGaussianSigmaPRF',
     'GaussianPRF',
     'GaussianPSF',
-    'IntegratedGaussianPRF',
     'MoffatPSF',
 ]
 
@@ -1480,105 +1478,6 @@ class CircularGaussianSigmaPRF(Fittable2DModel):
                 'y_0': inputs_unit[self.inputs[0]],
                 'sigma': inputs_unit[self.inputs[0]],
                 'flux': outputs_unit[self.outputs[0]]}
-
-
-@deprecated('2.0.0', alternative='`CircularGaussianSigmaPRF` or '
-            '`CircularGaussianPRF`')
-class IntegratedGaussianPRF(CircularGaussianSigmaPRF):
-    r"""
-    A circular 2D Gaussian PSF model integrated over pixels.
-
-    This model is evaluated by integrating the 2D Gaussian over the
-    input coordinate pixels, and is equivalent to assuming the PSF is
-    2D Gaussian at a *sub-pixel* level. Because it is integrated over
-    pixels, this model is considered a PRF instead of a PSF.
-
-    The Gaussian is normalized such that the analytical integral over
-    the entire 2D plane is equal to the total flux.
-
-    This model is equivalent to `CircularGaussianPRF`, but it is
-    parameterized in terms of the standard deviation (sigma) instead of
-    the full width at half maximum (FWHM).
-
-    Parameters
-    ----------
-    flux : float, optional
-        Total integrated flux over the entire PSF.
-
-    x_0 : float, optional
-        Position of the peak in x direction.
-
-    y_0 : float, optional
-        Position of the peak in y direction.
-
-    sigma : float, optional
-        Width of the Gaussian PSF.
-
-    bbox_factor : float, optional
-        The multiple of the standard deviation (sigma) used to define
-        the bounding box limits.
-
-    **kwargs : dict, optional
-        Additional optional keyword arguments to be passed to the
-        `astropy.modeling.Model` parent class.
-
-    See Also
-    --------
-    GaussianPSF, GaussianPRF, CircularGaussianPSF, CircularGaussianPRF
-
-    Notes
-    -----
-    The circular Gaussian function is defined as:
-
-    .. math::
-
-        f(x, y) =
-            \frac{F}{4}
-            \left[
-                {\rm erf} \left(\frac{x - x_0 + 0.5}
-                                     {\sqrt{2} \sigma} \right)  -
-                {\rm erf} \left(\frac{x - x_0 - 0.5}
-                                     {\sqrt{2} \sigma} \right)
-            \right]
-            \left[
-                {\rm erf} \left(\frac{y - y_0 + 0.5}
-                                     {\sqrt{2} \sigma} \right) -
-                {\rm erf} \left(\frac{y - y_0 - 0.5}
-                                     {\sqrt{2} \sigma} \right)
-            \right]
-
-    where :math:`F` is the total integrated flux, :math:`(x_{0},
-    y_{0})` is the position of the peak, :math:`\sigma` is the standard
-    deviation of the Gaussian, and :math:`{\rm erf}` denotes the error
-    function.
-
-    The model is normalized such that:
-
-    .. math::
-
-        \int_{-\infty}^{\infty} \int_{-\infty}^{\infty} f(x, y) \,dx \,dy = F
-
-    References
-    ----------
-    .. [1] https://en.wikipedia.org/wiki/Gaussian_function
-    """
-
-    flux = Parameter(
-        default=1, description='Total integrated flux over the entire PSF.')
-    x_0 = Parameter(
-        default=0, description='Position of the peak along the x axis')
-    y_0 = Parameter(
-        default=0, description='Position of the peak along the y axis')
-    sigma = Parameter(
-        default=1,
-        bounds=(FLOAT_EPSILON, None),
-        fixed=True,
-        description='Sigma (standard deviation) of the Gaussian')
-
-    def __init__(self, *, flux=flux.default, x_0=x_0.default, y_0=y_0.default,
-                 sigma=sigma.default, bbox_factor=5.5, **kwargs):
-        super().__init__(sigma=sigma, x_0=x_0, y_0=y_0, flux=flux,
-                         bbox_factor=bbox_factor, **kwargs)
 
 
 class MoffatPSF(Fittable2DModel):